Move create-flag to lib folder
