+++ packages/feature-flags/index.ts
@@ -1,3 +1,3 @@
-import { createFlag } from './create-flag';
+import { createFlag } from './lib/create-flag';
 
 export const showBetaFeature = createFlag('showBetaFeature');
